Moving github worflow parsing to its own file
