+++ checks/permissions.go
@@ -21,6 +21,7 @@ import (
 	"gopkg.in/yaml.v2"
 
 	"github.com/ossf/scorecard/v3/checker"
+	"github.com/ossf/scorecard/v3/checks/fileparser"
 	sce "github.com/ossf/scorecard/v3/errors"
 )
 
@@ -360,7 +361,7 @@ func testValidateGitHubActionTokenPermissions(pathfn string,
 // Check file content.
 func validateGitHubActionTokenPermissions(path string, content []byte,
 	dl checker.DetailLogger, data FileCbData) (bool, error) {
-	if !isWorkflowFile(path) {
+	if !fileparser.IsWorkflowFile(path) {
 		return true, nil
 	}
 	// Verify the type of the data.
