+++ checks/pinned_dependencies.go
@@ -16,7 +16,6 @@ package checks
 
 import (
 	"fmt"
-	"path"
 	"regexp"
 	"strings"
 
@@ -24,99 +23,18 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/ossf/scorecard/v3/checker"
+	"github.com/ossf/scorecard/v3/checks/fileparser"
 	sce "github.com/ossf/scorecard/v3/errors"
 )
 
 // CheckPinnedDependencies is the registered name for FrozenDeps.
 const CheckPinnedDependencies = "Pinned-Dependencies"
 
-// defaultShellNonWindows is the default shell used for GitHub workflow actions for Linux and Mac.
-const defaultShellNonWindows = "bash"
-
-// defaultShellWindows is the default shell used for GitHub workflow actions for Windows.
-const defaultShellWindows = "pwsh"
-
-// Structure for workflow config.
-// We only declare the fields we need.
-// Github workflows format: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions
-type gitHubActionWorkflowConfig struct {
-	Jobs map[string]gitHubActionWorkflowJob
-	Name string `yaml:"name"`
-}
-
-// A Github Action Workflow Job.
-// We only declare the fields we need.
-// Github workflows format: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions
-// nolint: govet
-type gitHubActionWorkflowJob struct {
-	Name     string                     `yaml:"name"`
-	Steps    []gitHubActionWorkflowStep `yaml:"steps"`
-	Defaults struct {
-		Run struct {
-			Shell string `yaml:"shell"`
-		} `yaml:"run"`
-	} `yaml:"defaults"`
-	RunsOn   stringOrSlice `yaml:"runs-on"`
-	Strategy struct {
-		// In most cases, the 'matrix' field will have a key of 'os' which is an array of strings, but there are
-		// some repos that have something like: 'matrix: ${{ fromJson(needs.matrix.outputs.latest) }}'.
-		Matrix interface{} `yaml:"matrix"`
-	} `yaml:"strategy"`
-}
-
-// A Github Action Workflow Step.
-// We only declare the fields we need.
-// Github workflows format: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions
-type gitHubActionWorkflowStep struct {
-	Name  string         `yaml:"name"`
-	ID    string         `yaml:"id"`
-	Shell string         `yaml:"shell"`
-	Run   string         `yaml:"run"`
-	If    string         `yaml:"if"`
-	Uses  stringWithLine `yaml:"uses"`
-}
-
-// stringOrSlice is for fields that can be a single string or a slice of strings. If the field is a single string,
-// this value will be a slice with a single string item.
-type stringOrSlice []string
-
-func (s *stringOrSlice) UnmarshalYAML(value *yaml.Node) error {
-	var stringSlice []string
-	err := value.Decode(&stringSlice)
-	if err == nil {
-		*s = stringSlice
-		return nil
-	}
-	var single string
-	err = value.Decode(&single)
-	if err != nil {
-		return sce.WithMessage(sce.ErrScorecardInternal, fmt.Sprintf("error decoding stringOrSlice Value: %v", err))
-	}
-	*s = []string{single}
-	return nil
-}
-
-// stringWithLine is for when you want to keep track of the line number that the string came from.
-type stringWithLine struct {
-	Value string
-	Line  int
-}
-
 // Structure to host information about pinned github
 // or third party dependencies.
 type worklowPinningResult struct {
 	thirdParties pinnedResult
 	gitHubOwned  pinnedResult
-}
-
-func (ws *stringWithLine) UnmarshalYAML(value *yaml.Node) error {
-	err := value.Decode(&ws.Value)
-	if err != nil {
-		return sce.WithMessage(sce.ErrScorecardInternal, fmt.Sprintf("error decoding stringWithLine Value: %v", err))
-	}
-	ws.Line = value.Line
-
-	return nil
 }
 
 //nolint:gochecknoinits
@@ -522,7 +440,7 @@ func testValidateGitHubWorkflowScriptFreeOfInsecureDownloads(pathfn string,
 // Returns true if the check should continue executing after this file.
 func validateGitHubWorkflowIsFreeOfInsecureDownloads(pathfn string, content []byte,
 	dl checker.DetailLogger, data FileCbData) (bool, error) {
-	if !isWorkflowFile(pathfn) {
+	if !fileparser.IsWorkflowFile(pathfn) {
 		return true, nil
 	}
 
@@ -533,7 +451,7 @@ func validateGitHubWorkflowIsFreeOfInsecureDownloads(pathfn string, content []by
 		return true, nil
 	}
 
-	var workflow gitHubActionWorkflowConfig
+	var workflow fileparser.GitHubActionWorkflowConfig
 	err := yaml.Unmarshal(content, &workflow)
 	if err != nil {
 		return false, sce.WithMessage(sce.ErrScorecardInternal,
@@ -552,7 +470,7 @@ func validateGitHubWorkflowIsFreeOfInsecureDownloads(pathfn string, content []by
 			}
 
 			// https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstepsrun.
-			shell, err := getShellForStep(&step, &job)
+			shell, err := fileparser.GetShellForStep(&step, &job)
 			if err != nil {
 				return false, err
 			}
@@ -577,99 +495,6 @@ func validateGitHubWorkflowIsFreeOfInsecureDownloads(pathfn string, content []by
 
 	addPinnedResult(pdata, validated)
 	return true, nil
-}
-
-// Returns the OSes this job runs on.
-func getOSesForJob(job *gitHubActionWorkflowJob) ([]string, error) {
-	// The 'runs-on' field either lists the OS'es directly, or it can have an expression '${{ matrix.os }}' which
-	// is where the OS'es are actually listed.
-	getFromMatrix := len(job.RunsOn) == 1 && strings.Contains(job.RunsOn[0], "matrix.os")
-	if !getFromMatrix {
-		return job.RunsOn, nil
-	}
-	jobOSes := make([]string, 0)
-	// nolint: nestif
-	if m, ok := job.Strategy.Matrix.(map[string]interface{}); ok {
-		if osVal, ok := m["os"]; ok {
-			if oses, ok := osVal.([]interface{}); ok {
-				for _, os := range oses {
-					if strVal, ok := os.(string); ok {
-						jobOSes = append(jobOSes, strVal)
-					}
-				}
-				return jobOSes, nil
-			}
-		}
-	}
-	return jobOSes, sce.WithMessage(sce.ErrScorecardInternal,
-		fmt.Sprintf("unable to determine OS for job: %v", job.Name))
-}
-
-// The only OS that this job runs on is Windows.
-func jobAlwaysRunsOnWindows(job *gitHubActionWorkflowJob) (bool, error) {
-	jobOSes, err := getOSesForJob(job)
-	if err != nil {
-		return false, err
-	}
-	for _, os := range jobOSes {
-		if !strings.HasPrefix(strings.ToLower(os), "windows") {
-			return false, nil
-		}
-	}
-	return true, nil
-}
-
-// getShellForStep returns the shell that is used to run the given step.
-func getShellForStep(step *gitHubActionWorkflowStep, job *gitHubActionWorkflowJob) (string, error) {
-	// https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#using-a-specific-shell.
-	if step.Shell != "" {
-		return step.Shell, nil
-	}
-	if job.Defaults.Run.Shell != "" {
-		return job.Defaults.Run.Shell, nil
-	}
-
-	isStepWindows, err := isStepWindows(step)
-	if err != nil {
-		return "", err
-	}
-	if isStepWindows {
-		return defaultShellWindows, nil
-	}
-
-	alwaysRunsOnWindows, err := jobAlwaysRunsOnWindows(job)
-	if err != nil {
-		return "", err
-	}
-	if alwaysRunsOnWindows {
-		return defaultShellWindows, nil
-	}
-
-	return defaultShellNonWindows, nil
-}
-
-// isStepWindows returns true if the step will be run on Windows.
-func isStepWindows(step *gitHubActionWorkflowStep) (bool, error) {
-	windowsRegexes := []string{
-		// Looking for "if: runner.os == 'Windows'" (and variants)
-		`(?i)runner\.os\s*==\s*['"]windows['"]`,
-		// Looking for "if: ${{ startsWith(runner.os, 'Windows') }}" (and variants)
-		`(?i)\$\{\{\s*startsWith\(runner\.os,\s*['"]windows['"]\)`,
-		// Looking for "if: matrix.os == 'windows-2019'" (and variants)
-		`(?i)matrix\.os\s*==\s*['"]windows-`,
-	}
-
-	for _, windowsRegex := range windowsRegexes {
-		matches, err := regexp.MatchString(windowsRegex, step.If)
-		if err != nil {
-			return false, sce.WithMessage(sce.ErrScorecardInternal, fmt.Sprintf("error matching Windows regex: %v", err))
-		}
-		if matches {
-			return true, nil
-		}
-	}
-
-	return false, nil
 }
 
 // Check pinning of github actions in workflows.
@@ -697,7 +522,7 @@ func testIsGitHubActionsWorkflowPinned(pathfn string, content []byte, dl checker
 // should continue executing after this file.
 func validateGitHubActionWorkflow(pathfn string, content []byte,
 	dl checker.DetailLogger, data FileCbData) (bool, error) {
-	if !isWorkflowFile(pathfn) {
+	if !fileparser.IsWorkflowFile(pathfn) {
 		return true, nil
 	}
 
@@ -709,7 +534,7 @@ func validateGitHubActionWorkflow(pathfn string, content []byte,
 		return true, nil
 	}
 
-	var workflow gitHubActionWorkflowConfig
+	var workflow fileparser.GitHubActionWorkflowConfig
 	err := yaml.Unmarshal(content, &workflow)
 	if err != nil {
 		return false, sce.WithMessage(sce.ErrScorecardInternal,
@@ -733,32 +558,12 @@ func validateGitHubActionWorkflow(pathfn string, content []byte,
 					})
 				}
 
-				githubOwned := isGitHubOwnedAction(step.Uses.Value)
-				addWorkflowPinnedResult(pdata, match, githubOwned)
-			}
+			githubOwned := fileparser.IsGitHubOwnedAction(execAction.Uses.Value)
+			addWorkflowPinnedResult(pdata, match, githubOwned)
 		}
 	}
 
 	return true, nil
-}
-
-// isWorkflowFile returns true if this is a GitHub workflow file.
-func isWorkflowFile(pathfn string) bool {
-	// From https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions:
-	// "Workflow files use YAML syntax, and must have either a .yml or .yaml file extension."
-	switch path.Ext(pathfn) {
-	case ".yml", ".yaml":
-		return true
-	default:
-		return false
-	}
-}
-
-// isGitHubOwnedAction check github specific action.
-func isGitHubOwnedAction(v string) bool {
-	a := strings.HasPrefix(v, "actions/")
-	c := strings.HasPrefix(v, "github/")
-	return a || c
 }
 
 func addWorkflowPinnedResult(w *worklowPinningResult, to, isGitHub bool) {
